Move the entry point to the module
