+++ pullmedown/musicforprogramming.py
@@ -3,6 +3,7 @@ import re
 from .utils import get_page, store
 import feedparser
 from collections import namedtuple
+from . import main
 
 MFPAlbum = namedtuple("MFPAlbum", "title link")
 
@@ -31,3 +32,23 @@ class MusicForProgrammingDownloader(object):
         print("{0} albums has been released on musicforprogramming".format(len(episode_list)))
         for ep in episode_list:
             print(ep.title)
+
+MFPDownloader = MusicForProgrammingDownloader
+
+@main.command()
+@click.option("--list-albums", is_flag=True, help="lists all the albums")
+@click.option("--number", default=0, help="download the specified song")
+@click.option("--download-all", is_flag=True, help="use it if you want to download all the albums")
+def musicforprogramming(list_albums, number, download_all):
+    if list_albums:
+        MFPDownloader().list_albums()
+        return
+    #raise NotImplementedError("you still need to wait for --number and --download-all")
+    if (number!=0) and download_all:
+        print("asking for all albums and a single one makes no sense")
+        return
+
+    if download_all:
+        MFPDownloader().download_all()
+    else:
+        MFPDownloader().download_episode(number)
